refactor: UI 테스트 파일 패키지 변경
- `app/src/androidTest/java/woowacourse/payments/ui/cardwallet/` 경로에 있던 테스트 파일을 `component` 하위 패키지로 이동
- `app/src/androidTest/java/woowacourse/payments/ui/newcard/` 경로에 있던 테스트 파일을 `component` 하위 패키지로 이동
- `PaymentCardTest` 및 `CardWalletContentTest`에서 `CardUiModel` 생성 시 `bankType` 파라미터 추가
